Completed most of the draft functionality
Users are now limited to ten position selections, and a running total of their total score is kept. Once a position is selected, a specific player performance is provided, and placed in the correct position once the user presses continue. The only position which does not work correctly is DST, as there are no photos to retreive for those 'players'. I created a team_photos folder to hold team logos which will go in place of a 'player photo'. However, the funcitonality for selecting a defense is still in progress.